+++ core/api_container/api/kurtosis_api.go
@@ -1,34 +1,55 @@
 package api
 
 import (
+	"context"
 	"github.com/docker/go-connections/nat"
 	"github.com/kurtosis-tech/kurtosis/commons/docker"
 	"github.com/kurtosis-tech/kurtosis/commons/networks"
 	"github.com/palantir/stacktrace"
 	"github.com/sirupsen/logrus"
 	"net/http"
+	"sync"
+	"time"
 )
 
 type KurtosisAPI struct {
+	// The Docker container ID of the test suite that will be making calls against this Kurtosis API
+	// This is expected to be nil until the "register test suite container" endpoint is called, which should be called
+	//  exactly once
+	testSuiteContainerId string
+
 	dockerManager *docker.DockerManager
 	dockerNetworkId string
 	freeIpAddrTracker *networks.FreeIpAddrTracker
 
-	// The Docker container ID of the test suite that will be making calls against this Kurtosis API
-	// This is expected to be nil until the "register test suite container" endpoint is called, which should be called
-	//  exactly once
-	testSuiteContainerId *string
+	// A value will be pushed to this channel iff a test execution has been registered with the API container and the
+	//  execution either completes successfully (as indicated by the testsuite container exiting) or hits the timeout
+	//  that was registered with the API container
+	testEndedBeforeTimeoutChan chan bool
+
+	// The names of the tests inside the suite; will be nil if no test suite has been registered yet
+	suiteTestNames []string
+
+	// Flag that will only be switched to true once, indicating that a test execution has been registered
+	testExecutionRegistered bool
+
+	mutex *sync.Mutex
 }
 
 func NewKurtosisAPI(
+		testSuiteContainerId string,
+		testEndedBeforeTimeoutChan chan bool,
 		dockerManager *docker.DockerManager,
 		dockerNetworkId string,
 		freeIpAddrTracker *networks.FreeIpAddrTracker) *KurtosisAPI {
 	return &KurtosisAPI{
-		dockerManager:        dockerManager,
-		dockerNetworkId:      dockerNetworkId,
-		freeIpAddrTracker:    freeIpAddrTracker,
-		testSuiteContainerId: nil,
+		testSuiteContainerId: testSuiteContainerId,
+		testEndedBeforeTimeoutChan: testEndedBeforeTimeoutChan,
+		dockerManager:              dockerManager,
+		dockerNetworkId:            dockerNetworkId,
+		freeIpAddrTracker:          freeIpAddrTracker,
+		suiteTestNames:             nil,
+		mutex:                      &sync.Mutex{},
 	}
 }
 
@@ -72,11 +93,96 @@ func (api *KurtosisAPI) StartService(httpReq *http.Request, args *StartServiceAr
 	return nil
 }
 
-func (api *KurtosisAPI) RegisterTestSuite(httpReq *http.Request, args *RegisterTestSuiteContainerArgs, result *struct{}) error {
-	// TODO kick off a thread to trigger the hard test timeout (prob with a Go channel)
+// Registers a test suite container with the API container, and throws an error if a test suite container has already
+//  been registered
+func (api *KurtosisAPI) RegisterTestSuite(httpReq *http.Request, args *RegisterTestSuiteArgs, result *struct{}) error {
+	api.mutex.Lock()
+	defer api.mutex.Unlock()
 
-	// TODO validation on container ID
-	// Strings are read-only so it's okay to do this
-	api.testSuiteContainerId = &(args.ContainerId)
+	if api.suiteTestNames != nil {
+		return stacktrace.NewError("A test suite is already registered to this API container")
+	}
+
+	testNames := args.TestNames
+	testNamesCopy := make([]string, len(testNames))
+	copy(testNamesCopy, testNames)
+
+	api.suiteTestNames = testNamesCopy
+
 	return nil
+}
+
+// Returns the names of the tests with the test suite registered with the API container, or an error if no test suite
+//  has been registered yet
+func (api *KurtosisAPI) ListTests(httpReq *http.Request, args *struct{}, result *ListTestsResponse) error {
+	api.mutex.Lock()
+	defer api.mutex.Unlock()
+
+	if api.suiteTestNames == nil {
+		return stacktrace.NewError("No test suite has been registered with this API container")
+	}
+
+	testNames := api.suiteTestNames
+	testNamesCopy := make([]string, len(testNames))
+	copy(testNamesCopy, testNames)
+	result.Tests = testNamesCopy
+
+	return nil
+}
+
+// Registers that the test suite container is going to run a test, and the Kurtosis API container should wait for the
+//  given amount of time before calling the test lost
+func (api *KurtosisAPI) RegisterTestExecution(httpReq *http.Request, args *RegisterTestExecutionArgs, result *struct{}) error {
+	api.mutex.Lock()
+	defer api.mutex.Unlock()
+
+	if api.testExecutionRegistered {
+		return stacktrace.NewError("A test execution is already registered with the API container")
+	}
+	api.testExecutionRegistered = true
+	go awaitTestCompletionOrTimeout(api.dockerManager, api.testSuiteContainerId, args.TestTimeoutSeconds, api.testEndedBeforeTimeoutChan)
+	return nil
+}
+
+/*
+Waits for either a) the testsuite container to exit or b) the given timeout to be reached, and pushes a corresponding
+	boolean value to the given channel based on which condition was hit
+ */
+func awaitTestCompletionOrTimeout(dockerManager *docker.DockerManager, testSuiteContainerId string, timeoutSeconds int, testEndedBeforeTimeoutChan chan bool) {
+	context, cancelFunc := context.WithCancel(context.Background())
+	defer cancelFunc()
+
+	// Kick off a thread that will only exit upon a) the testsuite contianer exiting or b) the context getting cancelled
+	testSuiteContainerExitedChan := make(chan struct{})
+	go awaitTestSuiteContainerExit(context, dockerManager, testSuiteContainerId, testSuiteContainerExitedChan)
+
+	timeout := time.Duration(timeoutSeconds) * time.Second
+	select {
+	case <- testSuiteContainerExitedChan:
+		// Triggered when the channel is closed, which is our signal that the testsuite container exited
+		testEndedBeforeTimeoutChan <- true
+	case <- time.After(timeout):
+		testEndedBeforeTimeoutChan <- false
+		cancelFunc() // We hit the timeout, so tell the container-awaiting thread to hara-kiri
+	}
+	close(testEndedBeforeTimeoutChan)
+}
+
+/*
+Waits for the container to exit until the context is cancelled
+ */
+func awaitTestSuiteContainerExit(
+		context context.Context,
+		dockerManager *docker.DockerManager,
+		testSuiteContainerId string,
+		testSuiteContainerExitedChan chan struct{}) {
+	_, err := dockerManager.WaitForExit(context, testSuiteContainerId)
+
+	if err != nil {
+		logrus.Debug("Got an error while waiting for the testsuite container to exit, likely indicating that the timeout was hit and the context was cancelled")
+	}
+
+	// If we get to here before the timeout, this will signal that the testsuite container exited; if we get to here
+	//  after the timeout, this won't do anything because nobody will be monitoring the other end of the channel
+	close(testSuiteContainerExitedChan)
 }