+++ core/CHANGELOG.md
@@ -2,6 +2,7 @@
 * Add a tutorial explaining what Kurtosis does at the Docker level
 * Kill TODOs in "Debugging Failed Tests" tutorial
 * Build a v0 of Docker container containing the Kurtosis API 
+* Add registration endpoint to the API container
 
 # 0.9.0
 * Change ConfigurationID to be a string